refactor: TotalSearchResponse
